requirements.txt
tqdm==4.63.0
atlassian-python-api==3.28.1
selenium==4.1.3
requests==2.27.1